requirements.txt
aniso8601==9.0.1
ansi2html==1.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backoff==1.11.1
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.8
bokeh==2.4.3
boto3==1.16.34
botocore==1.19.63
Brotli==1.0.9
cachetools==5.2.0
catalogue==2.0.7
catboost==1.0.6
category-encoders==2.5.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
chart-studio==1.1.0
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.5
colour==0.1.5
ConfigSpace==0.4.21
convertdate==2.4.0
cramjam==2.5.0
cryptography==37.0.4
cycler==0.11.0
cymem==2.0.6
Cython==0.29.30
daal==2021.6.0
daal4py==2021.6.3
dash==2.5.1
dash-auth==1.4.1
dash-bootstrap-components==1.2.0
dash-colorscales==0.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2021.11.2
dask-glm==0.2.0
dask-ml==2022.5.27
deap==1.3.1
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
distributed==2021.11.2
distro==1.7.0
docker==4.4.4
dpcpp-cpp-rt==2022.1.0
dtale==2.6.0
dtreeviz==1.3.6
emcee==3.1.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastai==2.5.6
fastcore==1.4.5
fastdownload==0.0.6
fastjsonschema==2.15.3
fastparquet==0.8.1
fastprogress==1.0.2
featuretools==1.11.1
Flask==2.1.2
Flask-Compress==1.12
flask-ngrok==0.0.25
Flask-RESTful==0.3.9
Flask-SeaSurf==1.1.1
flask-simplelogin==0.1.1
Flask-WTF==0.15.1
fonttools==4.33.3
fsspec==2022.5.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
graphviz==0.20
HeapDict==1.0.1
hijri-converter==2.2.4
holidays==0.14.2
idna==3.3
imageio==2.19.3
imbalanced-learn==0.9.1
impi-rt==2021.6.0
importlib-metadata==4.12.0
iniconfig==1.1.1
intel-cmplr-lib-rt==2022.1.0
intel-cmplr-lic-rt==2022.1.0
intel-fortran-rt==2022.1.0
intel-numpy==1.21.4
intel-opencl-rt==2022.1.0
intel-openmp==2022.1.0
intel-scipy==1.7.3
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-dash==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kaleido==0.2.1
kiwisolver==1.4.3
korean-lunar-calendar==0.2.1
kubernetes==12.0.1
langcodes==3.3.0
liac-arff==2.5.0
lightgbm==3.3.2
lime==0.2.0.1
llvmlite==0.38.1
locket==1.0.0
lz4==4.0.1
MarkupSafe==2.0.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
missingno==0.4.2
mistune==0.8.4
mkl==2022.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mkl-umath==0.1.1
msgpack==1.0.4
multipledispatch==0.6.0
murmurhash==1.0.7
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.5.1
neuraxle==0.7.0
nlp-primitives==2.7.1
nltk==3.7
notebook==6.4.12
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.0
openpyxl==3.0.10
oyaml==1.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathy==0.6.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
Pint==0.16.1
plotly==5.9.0
pluggy==1.0.0
pmdarima==1.8.5
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.12.0
PyMeeus==0.5.11
pymongo==4.1.1
pynisher==0.6.4
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypng==0.0.21
pyrfr==0.8.3
pyrsistent==0.18.1
pystache==0.6.0
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==23.2.0
regex==2022.7.9
requests==2.28.1
requests-oauthlib==1.3.1
requirements-parser==0.5.0
retrying==1.3.3
rsa==4.8
s3transfer==0.3.7
scikit-image==0.19.3
scikit-learn==1.1.1
scikit-learn-intelex==2021.6.3
scikit-MDR==0.4.4
scikit-optimize==0.9.0
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
shap==0.41.0
sigopt==8.5.0
six==1.16.0
sklearn-contrib-lightning==0.6.2.post0
skrebate==0.62
slicer==0.0.7
smac==1.2
smart-open==5.2.1
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
squarify==0.4.3
srsly==2.4.3
stack-data==0.3.0
statsmodels==0.13.2
stopit==1.1.2
strsimpy==0.2.1
tbb==2021.6.0
tbb4py==2021.6.3
tblib==1.7.0
tenacity==8.0.1
terminado==0.15.0
texttable==1.6.4
thinc==8.0.17
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
tomli==2.0.1
toolz==0.11.2
torch==1.11.0
torchvision==0.12.0
tornado==6.2
TPOT==0.11.7
tqdm==4.64.0
traitlets==5.3.0
typer==0.4.2
types-setuptools==62.6.0
typing_extensions==4.3.0
ua-parser==0.15.0
update-checker==0.18.0
urllib3==1.26.9
vowpalwabbit==9.1.0
waitress==2.1.2
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.1
woodwork==0.16.4
wrapt==1.14.1
WTForms==3.0.1
xarray==2022.3.0
xgboost==1.6.1
xlrd==2.0.1
zict==2.2.0
zipp==3.8.0
